ref(crons): Reorganize incident creation / issue occurrence logic (#80528)
Since we'll be doing more with issue occurrences split out the concept
of incidents into it's own logic module, as well as incident_occurrence
into it's own module

Part of GH-80527
