+++ src/sentry/monitors/types.py
@@ -102,6 +102,16 @@ class CheckinItem:
         )
 
 
+class SimpleCheckIn(TypedDict):
+    """
+    A stripped down check in object
+    """
+
+    id: int
+    date_added: datetime
+    status: int
+
+
 IntervalUnit = Literal["year", "month", "week", "day", "hour", "minute"]
